Refactor: Separate TriggerEffectHandler into smaller handlers
This commit refactors the `TriggerEffectHandler` by extracting logic for handling different effect types into separate handler objects: `ScoreHandlers`, `HealthHandlers`, `MultiplierHandlers`, `DamageHandlers`, and `MiscHandler`.
