+++ app/src/main/java/com/sq/thed_ck_licker/ecs/components/MultiplierComponent.kt
@@ -9,4 +9,10 @@ data class MultiplierComponent(var multiplier: Float = 1f): Component {
     fun removeMultiplier(multiplierToRemove: Float) {
         this.multiplier *= 1 / multiplierToRemove
     }
+
+    fun increaseMultiplier(multiplierToAdd: Float) {
+        this.multiplier += multiplierToAdd
+    }
+
+    fun decreaseMultiplier(multiplierToAdd: Float) = increaseMultiplier(-multiplierToAdd)
 }
